fixed errors

+++ E-commerce-back-End/order-service-api/src/main/java/icet/edu/dto/response/CustomerOrderResponse.java
@@ -1,8 +1,8 @@
 package icet.edu.dto.response;
 
-import icet.edu.dto.request.OrderDetailRequest;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -15,7 +15,7 @@ public class CustomerOrderResponse {
 
     private String orderId;
     private Date orderDate;
-    private Double totalAmount;
+    private BigDecimal totalAmount;
     private String userId;
     private String remark;
     private String status;
