+++ E-commerce-back-End/order-service-api/src/main/java/icet/edu/dto/response/OrderDetailResponse.java
@@ -2,6 +2,8 @@ package icet.edu.dto.response;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -12,7 +14,7 @@ public class OrderDetailResponse {
     private String detailId;
     private String productId;
     private int qty;
-    private double unitPrice;
-    private double discount;
+    private BigDecimal unitPrice;
+    private BigDecimal  discount;
 
 }
